requirements.txt
asgiref==3.11.0
Django==5.2.8
django-bootstrap5==25.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
model-bakery==1.20.5
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
redis==7.1.0
sqlparse==0.5.3
whitenoise==6.11.0
